update timetable movement controls

+++ store/states.ts
@@ -14,12 +14,8 @@ export const currentProjectState = atom<Irisub.Project | null>({
   effects_UNSTABLE: [
     ({ onSet, getPromise }) => {
       onSet(async (newValue, _oldValue, isReset: boolean) => {
-        if (!newValue) {
-          return;
-        }
-
         getPromise(databaseState).then((database) => {
-          if (database) {
+          if (database && newValue) {
             isReset
               ? database.delete(DB_STORES.PROJECT, newValue.id)
               : database.put(DB_STORES.PROJECT, newValue);
@@ -38,6 +34,17 @@ const currentTrackState = atom<string | null>({
 export const currentEventIndexState = atom<number>({
   key: "currentEventIndexState",
   default: 0,
+  effects_UNSTABLE: [
+    ({ onSet, setSelf, getPromise }) => {
+      onSet(async (newValue, oldValue) => {
+        if (newValue < 0) setSelf(oldValue);
+
+        const eventList = await getPromise(currentEventListState);
+        const lastIndex = eventList.length - 1;
+        if (newValue > lastIndex) setSelf(oldValue);
+      });
+    },
+  ],
 });
 
 export const currentEventListState = atom<Irisub.Event[]>({
